+++ example/coa.py
@@ -1,5 +1,4 @@
 #!/usr/bin/python
-from __future__ import print_function
 from pyrad.client import Client
 from pyrad import dictionary
 from pyrad import packet