+++ pyrad/client.py
@@ -8,7 +8,6 @@ import hashlib
 import select
 import socket
 import time
-import six
 import struct
 from pyrad import host
 from pyrad import packet
@@ -77,7 +76,7 @@ class Client(host.Host):
 
     def _SocketOpen(self):
         try:
-            family = socket.getaddrinfo(self.server, 'www')[0][0]
+            family = socket.getaddrinfo(self.server, 80)[0][0]
         except:
             family = socket.AF_INET
         if not self._socket: